Updated fill quad to draw the windows logo
