--- public/usage-examples/graphics/fill_quad_on_bitmap/fill_quad_on_bitmap-1-glass-window.txt
@@ -1,3 +1,0 @@
-#### Creating a Glass Window
-
-The following code demonstrates how to use [Fill Quad On Bitmap](/api/graphics#fill-quad-on-bitmap) to create a simple glass window using four window panels.
