+++ public/usage-examples/graphics/fill_quad_on_bitmap/fill_quad_on_bitmap-1-windows-logo.txt
@@ -1,0 +1,3 @@
+#### Creating a Glass Window
+
+The following code demonstrates how to use [Fill Quad On Bitmap](/api/graphics#fill-quad-on-bitmap) to create a simple windows logo using four window panels.
